# shared/contracts/openapi/fudcourt.yaml
openapi: 3.0.3
info:
  title: FUDCourt Executor API
  version: 0.1.0
  description: |
    The **executor HTTP surface** of FUDCourt as it exists TODAY.

    This contract documents the CURRENT TS-owned surface during the Go
    migration: the ground truth is the frozen contract in
    `frontend/web/src/platform/executor/types.ts` and the route handlers under
    `frontend/web/src/app/**/api/executor/**` (endpoint list mirrored from the
    header of `frontend/web/src/features/executor/client.ts`). It is **the
    compatibility contract for `backend/api` + `backend/workers/executor` +
    `shared/sdk/typescript`**: the Go services being built MUST preserve these
    methods, paths, and response envelopes, and `shared/sdk/typescript` generates its
    types from this file.

    House rules encoded here (PRD §108, §109, house style):

    * **Unknown fields are forbidden to be added by consumers.** Request bodies
      are closed (`additionalProperties: false`); do not extend this contract
      ad hoc — extend this document and regenerate the SDK.
    * **Honest nulls**: fields that may be unknown are `nullable`, never
      faked to `0`.
    * **Credentials are write-only**: secrets are accepted once at connect
      time, sealed server-side, and no route ever returns them. Only masked
      keys (`abc...xyz`) leave the server.
    * Error bodies keep the server's own text (`error`, `detail`, `errors`) —
      clients display `err.message` verbatim.

...
  license:
...
servers:
...
paths:
...
  /api/executor/accounts:
    get:
      tags: [accounts]
      operationId: listAccounts
      summary: Connected accounts — masked keys only (PRD §87, §109)
      responses:
        '200':
          description: Credential records (secrets are never returned).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccountsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [accounts]
      operationId: connectAccount
      summary: BYOK connect — secrets are write-only, sealed server-side
      description: |
        `apiKey`/`apiSecret`/`passphrase` are accepted once here and never
        returned by any route. The live venue probe runs at connect time; a key
        that grants withdrawal permission is refused (400).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectAccountInput'
      responses:
        '200':
          description: The sealed record (masked key) and the venue's probe metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid body, or `withdrawal permission not supported` / validation errors.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/ErrorBasic'
                  - $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/executor/accounts/{id}:
    get:
      tags: [accounts]
      operationId: getAccount
      summary: One connected account — masked key only
      description: |
        Present as a route handler today with no `client.ts` function yet; kept
        in the contract so the Go migration preserves it.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: The sealed record (masked key) and the venue's probe metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [accounts]
      operationId: deleteAccount
      summary: Revoke a credential (secrets destroyed; record kept masked)
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/executor/accounts/{id}/test:
    post:
      tags: [accounts]
      operationId: testAccount
      summary: Live credential validation (PRD §46)
      description: |
        Re-probes the venue with the sealed credential and refreshes its
        health. A failing probe answers `502 { error: 'credential check
        failed', detail, category }` — the health is still updated.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: The sealed record (masked key) and fresh probe metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: "`withdrawal permission not supported`."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
...
components:
  parameters:
...
    AccountId:
      name: id
      in: path
      required: true
      description: Credential record id.
      schema:
        type: string
  responses:
    Unauthorized:
      description: No session, or the session lacks the `team` tier (fail-closed).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    NotFound:
      description: Not found — or not yours (ownership is never leaked, PRD §108).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBasic'
    UpstreamFailure:
      description: Exchange/venue probe failed; `category` is the retry taxonomy (PRD §78).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
  schemas:
...
    ExchangeId:
      type: string
      description: 'Exchange identity. Keys in `exchange:*:*` venue keys (PRD §49).'
      enum:
        - binance
        - bybit
        - mexc
...
    CredentialHealth:
      type: string
      description: 'Credential health (PRD §47).'
      enum:
        - ACTIVE
        - INVALID
        - EXPIRED
        - PERMISSION_ERROR
        - RATE_LIMITED
        - REVOKED
        - UNKNOWN
    ErrorCategory:
      type: string
      description: 'Retry taxonomy (PRD §78). Only `network_retryable`/`rate_limited`/`exchange_overload` may retry.'
      enum:
        - network_retryable
        - rate_limited
        - exchange_overload
        - invalid_order
        - permission_error
        - insufficient_balance
        - fatal
        - unknown
...
    AccountPermissions:
      type: object
      additionalProperties: false
      required: [read, spotTrade, futuresTrade, withdraw]
      properties:
        read:
          type: boolean
          description: 'Verified by probe: the credential can read account state.'
        spotTrade:
          type: boolean
          nullable: true
          description: 'As REPORTED by the venue''s key-restriction endpoint where one exists; `null` = the venue does not report it — never inferred.'
        futuresTrade:
          type: boolean
          nullable: true
        withdraw:
          type: boolean
          nullable: true
          description: 'MUST be false or null in effect: FUDCourt NEVER requests or uses withdrawal capability (PRD §43). `true` renders as a UI warning.'
    AccountMetadata:
      type: object
      additionalProperties: true
      required: [exchange, label, accountType, permissions, health, apiKeyMasked]
      description: 'Venue probe result for a credential. Masked key only (PRD §109).'
      properties:
        exchange:
          $ref: '#/components/schemas/ExchangeId'
        label:
          type: string
          nullable: true
        accountType:
          type: string
          nullable: true
        permissions:
          $ref: '#/components/schemas/AccountPermissions'
        health:
          $ref: '#/components/schemas/CredentialHealth'
        apiKeyMasked:
          type: string
          nullable: true
          description: 'Partial key display only: `abc...xyz`. NEVER the plaintext key.'
        raw:
          description: Optional venue-specific probe detail (JSON round-trippable).
    CredentialRecord:
      type: object
      additionalProperties: false
      required: [id, userId, exchange, label, apiKeyMasked, permissions, health, createdAt, updatedAt, lastUsedAt, revokedAt]
      description: 'A sealed BYOK credential. Masked only — no route ever returns plaintext secrets (PRD §44, §108, §109).'
      properties:
        id:
          type: string
        userId:
          type: string
        exchange:
          $ref: '#/components/schemas/ExchangeId'
        label:
          type: string
        apiKeyMasked:
          type: string
          description: 'Masked `abc...xyz` only — never the full key (PRD §109).'
        permissions:
          $ref: '#/components/schemas/AccountPermissions'
        health:
          $ref: '#/components/schemas/CredentialHealth'
        createdAt:
          type: integer
          description: Unix milliseconds.
        updatedAt:
          type: integer
        lastUsedAt:
          type: integer
          nullable: true
        revokedAt:
          type: integer
          nullable: true
...
    ConnectAccountInput:
      type: object
      additionalProperties: false
      required: [exchange, label, apiKey, apiSecret]
      description: 'BYOK connect body. Secrets are write-only: no route ever reads them back.'
      properties:
        exchange:
          type: string
        label:
          type: string
        apiKey:
          type: string
          writeOnly: true
        apiSecret:
          type: string
          writeOnly: true
        passphrase:
          type: string
          writeOnly: true
...
    ListAccountsResponse:
      type: object
      additionalProperties: false
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/CredentialRecord'
    AccountResponse:
      type: object
      additionalProperties: false
      required: [account, metadata]
      description: Connect, test and get all answer the sealed record and the venue's probe result.
      properties:
        account:
          $ref: '#/components/schemas/CredentialRecord'
        metadata:
          $ref: '#/components/schemas/AccountMetadata'
    DeleteAccountResponse:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    ErrorBasic:
      type: object
      additionalProperties: true
      required: [error]
      description: 'Minimal route error — e.g. `{"error": "invalid JSON body"}` or `{"error": "account not found"}`.'
      properties:
        error:
          type: string
          description: The server's own error text; display verbatim.
    ErrorDetail:
      type: object
      additionalProperties: true
      required: [error, detail]
      description: 'Error with a specific reason — e.g. `{"error": "invalid transition", "detail": "..."}` or `{"error": "credential check failed", "detail": "...", "category": "..."}`. Extra numeric fields may carry the numbers behind a refusal (PRD §73/§74).'
      properties:
        error:
          type: string
        detail:
          type: string
        category:
          $ref: '#/components/schemas/ErrorCategory'
    ValidationError:
      type: object
      additionalProperties: true
      required: [error, errors]
      description: 'Field-named validation errors — `{"error": "validation", "errors": ["symbol: ...", ...]}`.'
      properties:
        error:
          type: string
          enum: [validation]
        errors:
          type: array
          items:
            type: string
...